LPS-106594 Move to ReleaseDataFactory
