+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/macro.ftl
@@ -13,6 +13,7 @@
 	layoutDataFactory = dataFactory.getDataFactoryInstance("layoutDataFactory")
 	messageBoardDataFactory = dataFactory.getDataFactoryInstance("messageBoardDataFactory")
 	portletPreferenceDataFactory = dataFactory.getDataFactoryInstance("portletPreferenceDataFactory")
+	releaseDataFactory = dataFactory.getDataFactoryInstance("releaseDataFactory")
 />
 
 <#macro insertAssetEntry
